requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
pandas==2.2.2
matplotlib==3.9.0
markdown==3.6
python-dotenv==1.0.1
feedparser>=6.0.11
python-dateutil>=2.9.0.post0
